App.jsからliff.initを除去
